将authorid改为Long
1

+++ src/main/java/cn/edu/buaa/scholarshipserver/dao/DataScholarDao.java
@@ -7,7 +7,7 @@ import org.springframework.data.elasticsearch.repository.ElasticsearchRepository
 import java.util.List;
 
 public interface DataScholarDao extends ElasticsearchRepository<DataScholar,Long> {
-    public DataScholar findByAuthorId(int authorId);
+    public DataScholar findByAuthorId(Long authorId);
     public List<DataScholar> findByNormalizedName(String normalizedName);
     public List<DataScholar> findByScholarId(Integer scholarId);
