+++ src/main/java/cn/edu/buaa/scholarshipserver/dao/PaperDataScholarDao.java
@@ -6,5 +6,5 @@ import org.springframework.data.elasticsearch.repository.ElasticsearchRepository
 import java.util.List;
 
 public interface PaperDataScholarDao extends ElasticsearchRepository<Paper_DataScholar,Long> {
-        public List<Paper_DataScholar> findByAuthorId(Integer authorId);
+        public List<Paper_DataScholar> findByAuthorId(Long authorId);
 }
